+++ cli/pack/app_dir.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/apex/log"
 	"github.com/otiai10/copy"
+	"gopkg.in/yaml.v2"
 
 	"github.com/tarantool/cartridge-cli/cli/build"
 	"github.com/tarantool/cartridge-cli/cli/common"
@@ -23,10 +24,23 @@ const (
 	versionFileName    = "VERSION"
 	versionLuaFileName = "VERSION.lua"
 
-	maxCachedProjects = 5
+	cacheParamsFileName     = "pack-cache-config.yml"
+	maxCachedProjects       = 5
+	cntFirstSymbolsFromHash = 10
+
+	cacheParamsErrorMsg = "Please, specify one and only one of `always-true`, `key` and `key-path` for path %s"
 )
 
 type CachePaths map[string]string
+
+type CachePathParams struct {
+	Path        string `yaml:"path,omitempty"`
+	Key         string `yaml:"key,omitempty"`
+	KeyPath     string `yaml:"key-path,omitempty"`
+	AlwaysCache bool   `yaml:"always-cache,omitempty"`
+}
+
+type CachePathsParams []CachePathParams
 
 func initAppDir(appDirPath string, ctx *context.Ctx) error {
 	var err error
@@ -55,9 +69,24 @@ func initAppDir(appDirPath string, ctx *context.Ctx) error {
 		return err
 	}
 
+	log.Debugf("Creating cache directory")
+	if err := os.MkdirAll(ctx.Cli.CacheDir, 0755); err != nil {
+		return fmt.Errorf("Failed to create cache directory: %s", err)
+	}
+
+	if _, err := os.Stat(filepath.Join(ctx.Project.Path, cacheParamsFileName)); err != nil {
+		// File exists, but we can't process it
+		if !os.IsNotExist(err) {
+			return fmt.Errorf("Failed to process %s file which contain cache paths", cacheParamsFileName)
+		}
+
+		// File doesn't exists and this is okay, we just ignoring cache
+		ctx.Pack.NoCache = true
+	}
+
 	cachePaths, err := getProjectCachePaths(ctx)
 	if err != nil {
-		log.Warnf("%s", err)
+		return fmt.Errorf("Failed to get cache paths: %s", err)
 	}
 
 	copyFromCache(cachePaths, appDirPath, ctx)
@@ -68,14 +97,14 @@ func initAppDir(appDirPath string, ctx *context.Ctx) error {
 		return err
 	}
 
-	// Update cache in cartridge temp directory
-	if err := updateCache(cachePaths, ctx); err != nil {
-		log.Warnf("%s", err)
-	}
-
 	// post-build
 	if err := build.PostRun(ctx); err != nil {
 		return err
+	}
+
+	// Update cache in cartridge temp directory
+	if err := updateCache(cachePaths, ctx); err != nil {
+		log.Warnf("%s", err)
 	}
 
 	// generate VERSION file
@@ -106,24 +135,101 @@ func copyFromCache(paths CachePaths, destPath string, ctx *context.Ctx) {
 
 	for path, cacheDir := range paths {
 		if _, err := os.Stat(cacheDir); err == nil {
-			if err := copyPathFromCache(cacheDir, filepath.Join(destPath, path)); err != nil {
+			if err := copyPathFromCache(cacheDir, filepath.Join(destPath, path), path); err != nil {
 				log.Warnf("%s", err)
 			}
 		} else if !os.IsNotExist(err) {
-			log.Warnf("Failed to copy from cache: %s", err)
+			log.Warnf("Failed to copy path %s from cache: %s", cacheDir, err)
 		}
 	}
 }
 
-func copyPathFromCache(cachedPath string, destPath string) error {
-	log.Infof("Using cached path %s", filepath.Base(destPath))
-	err := copy.Copy(cachedPath, destPath)
-
+func cachePathIsFile(cachePath string, baseDestPath string) (bool, error) {
+	files, err := ioutil.ReadDir(cachePath)
 	if err != nil {
+		return false, err
+	}
+
+	if len(files) == 1 {
+		fileName := files[0].Name()
+		if fileName == baseDestPath {
+			if fileInfo, err := os.Stat(filepath.Join(cachePath, fileName)); err == nil {
+				return !fileInfo.IsDir(), nil
+			}
+
+			return false, err
+		}
+	}
+
+	return false, nil
+}
+
+func copyPathFromCache(cachedPath string, destPath string, pathFromRoot string) error {
+	baseDestPath := filepath.Base(destPath)
+	cacheIsFile, err := cachePathIsFile(cachedPath, baseDestPath)
+	if err != nil {
+		return fmt.Errorf("Failed to determine if the path is a file: %s", err)
+	}
+
+	if cacheIsFile {
+		cachedPath = filepath.Join(cachedPath, baseDestPath)
+	}
+
+	if err := copy.Copy(cachedPath, destPath); err != nil {
 		return fmt.Errorf("Failed to copy path %s from cache to project directory: %s", destPath, err)
 	}
 
+	log.Infof("Using cached path %s", pathFromRoot)
 	return nil
+}
+
+func parseCacheParamsFile(cacheParamsPath string) (CachePathsParams, error) {
+	if _, err := os.Stat(cacheParamsPath); os.IsNotExist(err) {
+		return nil, fmt.Errorf("File %s with pack cache paths parameters doesn't exists", cacheParamsPath)
+	} else if err != nil {
+		return nil, fmt.Errorf("Failed to process %s file: %s", cacheParamsPath, err)
+	}
+
+	fileContent, err := common.GetFileContentBytes(cacheParamsPath)
+	if err != nil {
+		return nil, fmt.Errorf("Failed to read %s file: %s", cacheParamsPath, err)
+	}
+
+	var cachePathsParams CachePathsParams
+	if err := yaml.Unmarshal(fileContent, &cachePathsParams); err != nil {
+		return nil, fmt.Errorf("Failed to parse file %s with pack cache parameters: %s", cacheParamsPath, err)
+	}
+
+	return cachePathsParams, nil
+}
+
+func calculateCachePath(ctx *context.Ctx, params *CachePathParams) (string, error) {
+	var keyHash string
+	var err error
+
+	projectPathHash := common.StringSHA1Hex(ctx.Project.Path)[:cntFirstSymbolsFromHash]
+
+	switch {
+	case params.AlwaysCache == true:
+		keyHash = "always"
+	case params.KeyPath != "":
+		pathFromProjectRoot := filepath.Join(ctx.Project.Path, params.KeyPath)
+		if _, err := os.Stat(pathFromProjectRoot); err != nil {
+			return "", fmt.Errorf("Failed to get specified cache key file for path %s: %s", params.Path, err)
+		}
+
+		if keyHash, err = common.FileSHA1Hex(pathFromProjectRoot); err != nil {
+			return "", fmt.Errorf("Failed to get hash from file content for path %s: %s", params.Path, err)
+		}
+
+		keyHash = keyHash[:cntFirstSymbolsFromHash]
+	case params.Key != "":
+		keyHash = common.StringSHA1Hex(params.Key)[:cntFirstSymbolsFromHash]
+	default:
+		panic("Failed to calculate cache path: `key` and `key-path` fields are empty and `always-cache` flag is false")
+	}
+
+	return filepath.Join(ctx.Cli.CacheDir, projectPathHash, params.Path, keyHash), nil
 }
 
 func getProjectCachePaths(ctx *context.Ctx) (CachePaths, error) {
@@ -131,23 +237,30 @@ func getProjectCachePaths(ctx *context.Ctx) (CachePaths, error) {
 		return nil, nil
 	}
 
-	rockspecPath, err := common.FindRockspec(ctx.Project.Path)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to find rockspec: %s", err)
-	} else if rockspecPath == "" {
-		return nil, fmt.Errorf("Application directory should contain rockspec")
-	}
-
-	projectPathHash := common.StringSHA1Hex(ctx.Project.Path)[:10]
-	rockspecHash, err := common.FileSHA1Hex(rockspecPath)
+	cachePathsParams, err := parseCacheParamsFile(filepath.Join(ctx.Project.Path, cacheParamsFileName))
 	if err != nil {
 		return nil, err
 	}
 
-	rockspecHash = rockspecHash[:10]
+	cachePaths := CachePaths{}
+	for _, params := range cachePathsParams {
+		if !common.OnlyOneIsTrue(params.Key != "", params.KeyPath != "", params.AlwaysCache) {
+			return nil, fmt.Errorf(cacheParamsErrorMsg, params.Path)
+		}
 
-	return CachePaths{
-		".rocks": filepath.Join(ctx.Cli.CacheDir, projectPathHash, ".rocks", rockspecHash)}, nil
+		cachePath, err := calculateCachePath(ctx, &params)
+		if err != nil {
+			return nil, err
+		}
+
+		if _, found := cachePaths[params.Path]; found {
+			return nil, fmt.Errorf("Cache path %s specified multiple times", params.Path)
+		}
+
+		cachePaths[params.Path] = cachePath
+	}
+
+	return cachePaths, nil
 }
 
 func rotateCacheDirs(ctx *context.Ctx) error {
@@ -173,7 +286,7 @@ func rotateCacheDirs(ctx *context.Ctx) error {
 }
 
 func updateCache(paths CachePaths, ctx *context.Ctx) error {
-	if ctx.Pack.NoCache {
+	if ctx.Pack.NoCache || len(paths) == 0 {
 		return nil
 	}
 
@@ -189,8 +302,20 @@ func updateCache(paths CachePaths, ctx *context.Ctx) error {
 			log.Warnf("Failed to clear %s cache directory: %s", currentPath, err)
 		}
 
-		if err := copy.Copy(filepath.Join(ctx.Build.Dir, path), cacheDir); err != nil {
-			log.Warnf("Failed to copy %s from cache: %s", path, err)
+		copyPath := filepath.Join(ctx.Build.Dir, path)
+		if fileInfo, err := os.Stat(copyPath); err == nil {
+			if !fileInfo.IsDir() {
+				cacheDir = filepath.Join(cacheDir, filepath.Base(path))
+			}
+		}
+
+		// NOTE: This approach does not allow copying packages
+		// (like tar.gz, rpm or deb). Apparently for the same reason,
+		// temporary directory with application files do not contain
+		// such packages, because they were copied from the directory
+		// in the same way.
+		if err := copy.Copy(copyPath, cacheDir); err != nil {
+			log.Warnf("Failed copy %s to cache: %s", copyPath, err)
 		}
 
 		log.Debugf("%s cache has been successfully saved in: %s", path, cacheDir)